// DesktopApp/src/types/ClickUpTaskDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ClickUpAttachment } from "./ClickUpAttachment";
import type { ClickUpComment } from "./ClickUpComment";

export type ClickUpTaskDetail = { id: string, customId: string | null, name: string, description: string, status: string | null, url: string, attachments: Array<ClickUpAttachment>, comments: Array<ClickUpComment>, };